src/lambda/requirements.txt
# Core dependencies
strands-agents==0.1.6
strands-agents-tools==0.1.4
boto3>=1.38.0

# These dependencies are already included in the Lambda Python runtime
# but are listed here for completeness
json
logging
os
traceback
datetime
